Add files via upload
Weather Forecast Dashboard allows users to search for the weather of any city by entering the city name. It retrieves the current temperature, wind speed, and a 7-hour forecast using real-time data from the Open-Meteo and OpenWeatherMap APIs. The app features a modern design with smooth animations and interactive elements, making it easy and engaging to check weather conditions for different locations.

!!!Add your OpenWeatherMap API key in the JS file so it will work!!!